Convert cwftoflux to a proper flux source.
