+++ lib/fluxsource/fluxsource.h
@@ -3,14 +3,15 @@
 
 #include "flags.h"
 
-class Fluxmap;
-class FluxSpec;
-class FluxSourceProto;
-class HardwareFluxSourceProto;
-class TestPatternFluxSourceProto;
+class CwfFluxSourceProto;
 class EraseFluxSourceProto;
+class FluxSourceProto;
+class FluxSpec;
+class Fluxmap;
+class HardwareFluxSourceProto;
 class KryofluxFluxSourceProto;
 class ScpFluxSourceProto;
+class TestPatternFluxSourceProto;
 
 class FluxSource
 {
@@ -18,12 +19,13 @@ public:
     virtual ~FluxSource() {}
 
 private:
-    static std::unique_ptr<FluxSource> createSqliteFluxSource(const std::string& filename);
+    static std::unique_ptr<FluxSource> createCwfFluxSource(const CwfFluxSourceProto& config);
+    static std::unique_ptr<FluxSource> createEraseFluxSource(const EraseFluxSourceProto& config);
     static std::unique_ptr<FluxSource> createHardwareFluxSource(const HardwareFluxSourceProto& config);
     static std::unique_ptr<FluxSource> createKryofluxFluxSource(const KryofluxFluxSourceProto& config);
     static std::unique_ptr<FluxSource> createScpFluxSource(const ScpFluxSourceProto& config);
+    static std::unique_ptr<FluxSource> createSqliteFluxSource(const std::string& filename);
     static std::unique_ptr<FluxSource> createTestPatternFluxSource(const TestPatternFluxSourceProto& config);
-    static std::unique_ptr<FluxSource> createEraseFluxSource(const EraseFluxSourceProto& config);
 
 public:
     static std::unique_ptr<FluxSource> create(const FluxSourceProto& spec);
