+++ lib/fluxsource/fluxsource.proto
@@ -29,6 +29,11 @@ message ScpFluxSourceProto {
 		(help) = ".scp file to read flux from"];
 }
 
+message CwfFluxSourceProto {
+	optional string filename = 1 [default = "flux.xwf",
+		(help) = ".cwf file to read flux from"];
+}
+
 message FluxSourceProto {
 	oneof source {
 		string fluxfile = 1;
@@ -37,6 +42,7 @@ message FluxSourceProto {
 		EraseFluxSourceProto erase = 4;
 		KryofluxFluxSourceProto kryoflux = 5;
 		ScpFluxSourceProto scp = 6;
+		CwfFluxSourceProto cwf = 7;
 	}
 }
